Modifying delete product route to delete products only you added

+++ src/routers/products.js
@@ -7,7 +7,7 @@ const auth = require('../middleware/auth')
 const  router = new express.Router()
 
 //(Private) Get all products
-router.get('/products', async (req,res) =>{
+router.get('/products',auth ,async (req,res) =>{
     try{
         const products = await Product.find({})
         res.send(products)
@@ -17,7 +17,7 @@ router.get('/products', async (req,res) =>{
 }) 
 
 //(Private) Get product by it's id
-router.get('/products/:id', async (req,res) =>{
+router.get('/products/:id',auth ,async (req,res) =>{
     try{
         const product = await Product.findById(req.params.id)
         res.send(product)
@@ -42,7 +42,7 @@ router.post('/products/addItem' , auth, async (req,res) => {
 })
 
 //(Private) purchasing a product
-router.post('/products/purchaseItem/:id' ,  async (req,res) => {
+router.post('/products/purchaseItem/:id' , auth , async (req,res) => {
     try{
         const _id = req.params.id
         const product = await Product.findOne({_id})
@@ -64,14 +64,25 @@ router.post('/products/purchaseItem/:id' ,  async (req,res) => {
 })
 
 //(Private) delete a product  
-router.delete('/products/deleteItem/:id' , async  (req,res)=>{
+router.delete('/products/deleteItem/:id', auth ,async  (req,res)=>{
     try{
-        await Product.deleteOne({_id : req.params.id})
-        req.user.inventory = req.user.inventory.filter((elem) =>{
-            return elem._id.toString() !== req.params.id
-        })
-        await req.user.save()
-        res.status(200).send(req.user)
+        let valid_delete = 0 
+        for (let i = 0 ; i < req.user.inventory.length ; i++){
+            console.log(req.user.inventory[i]._id.toString() , req.params.id)
+            if (req.user.inventory[i]._id.toString() === req.params.id){
+                valid_delete = 1
+            }
+        }
+        if (valid_delete){
+            await Product.deleteOne({_id : req.params.id})
+            req.user.inventory = req.user.inventory.filter((elem) =>{
+                return elem._id.toString() !== req.params.id
+            })
+            await req.user.save()
+            res.status(200).send(req.user)
+        } else {
+            throw new Error ('Cannot delete item.')
+        }
     } catch (error){
         res.status(404).send(error.message)
     }
